requirements.txt
pelican>=3.6.0
pynacl>=1.0.0
